LDPMLab

Documentation for LDPMLab.

LDPMLab.Boundary_setting

LDPMLab.Particle_distribution

Hello

\[1+1 = A\]

[1,1]+[0.2,0.4]

2-element Vector{Float64}:
 1.2
 1.4

Hello

LDPMLab.Boundary_setting — Function

eigenbox(A[, method=Rohn()])

Returns an enclosure of all the eigenvalues of A. If A is symmetric, then the output is a real interval, otherwise it is a complex interval.

Input

A – square interval matrix

method – method used to solve the symmetric interval eigenvalue problem (bounding eigenvalues of general matrices is also reduced to the symmetric case). Possible values are

- `Rohn` -- (default) fast method to compute an enclosure of the eigenvalues of
      a symmetric interval matrix
- `Hertz` -- finds the exact hull of the eigenvalues of a symmetric interval
      matrix, but has exponential complexity.

Algorithm

LDPMLab.Particle_distribution — Function

particle_distribution(model_name, particle_save="Yes", particle_dirc_and_name="LDPM_particle_distribution")

Excute particle distribution in material volume and save the data of particle distribution.

Input

model_name – symbol describing the model type used

:LDPM – uses LDPM

:LDPM_bar_reforced – uses rank1 update

particle_save – symbol describing wether saving particle distribution

:Yes – save a JLD2 file with particle coordinates and their corresponding diameters. By default, :Yes is used. Other strings except Yes means not storing the current particle distribution.

Saving particle distribution is encouraged so that the following solutions can use the same model mesh, excluding the influence of meshing on solutions.

particle_dirc_and_name – a string indicating the filefolder and filename saving particle distribution. By default, LDPM_particle_distribution is used. This argument can be ../output examples/Juliafiles/LDPM_particle_distribution as well.